+++ private/buf/bufctl/controller.go
@@ -691,11 +691,11 @@ func (c *controller) GetMessage(
 	}
 	var validator protoyaml.Validator
 	if functionOptions.messageValidation {
-		var err error
-		validator, err = protovalidate.New()
+		protovalidateValidator, err := protovalidate.New()
 		if err != nil {
 			return nil, 0, err
 		}
+		validator = yamlValidator{protovalidateValidator}
 	}
 	var unmarshaler protoencoding.Unmarshaler
 	switch messageEncoding {
@@ -1285,6 +1285,15 @@ func (c *controller) handleFileAnnotationSetRetError(retErrAddr *error) {
 		}
 		*retErrAddr = ErrFileAnnotation
 	}
+}
+
+// Implements [protoyaml.Validator] using [protovalidate.Validator]. This allows us to pass
+// a [protovalidate.Validator] to the [protoencoding.NewYAMLUnmarshaler] and validate types
+// when unmarshalling.
+type yamlValidator struct{ protovalidateValidator protovalidate.Validator }
+
+func (v yamlValidator) Validate(msg proto.Message) error {
+	return v.protovalidateValidator.Validate(msg)
 }
 
 func getImageFileInfosForModuleSet(ctx context.Context, moduleSet bufmodule.ModuleSet) ([]bufimage.ImageFileInfo, error) {